requirements.txt
# Core Dependencies
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Deep Learning
ultralytics==8.1.0
torch==2.1.0
torchvision==0.16.0

# TensorRT (install separately on Jetson)
# tensorrt>=8.5.0

# OCR
easyocr==1.7.1
Pillow==10.1.0

# Cloud Database
firebase-admin==6.3.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
